Move submit_net logic into Backend
Now that the celery app uses the Factory, submit_net logic can move
into the Backend.
